chore(registration-react): Rename redux module "auth" to "user"
