+++ applications/registration-react/src/index.jsx
@@ -9,7 +9,7 @@ import { configureLocalization } from './lib/localization';
 import { App } from './app/app';
 import { configureReferenceDataApi } from './api/reference-data-api';
 import { configureRegistrationApi } from './api/registration-api';
-import { getUserProfileThunk } from './redux/modules/auth';
+import { getUserProfileThunk } from './redux/modules/user';
 
 import './styles';
