+++ applications/registration-react/src/redux/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
-import auth from './modules/auth';
+import user from './modules/user';
 import catalog from './modules/catalog';
 import datasets from './modules/datasets';
 import catalogs from './modules/catalogs';
@@ -12,7 +12,7 @@ import { referenceDataReducer } from './modules/referenceData';
 
 const rootReducer = combineReducers({
   form: formReducer,
-  auth,
+  user,
   catalog,
   datasets,
   catalogs,
